🚀 US-36 Alteração para Long nos Repository do User e UserSpaceRole

+++ us-twogether/src/main/java/com/app/us_twogether/repository/SpaceRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface SpaceRepository extends JpaRepository<Space, Integer> {
+public interface SpaceRepository extends JpaRepository<Space, Long> {
     Optional<Space> findBySharedToken(String sharedToken);
     @Query("SELECT spc FROM Space spc JOIN UserSpaceRole usr ON spc = usr.space WHERE usr.user = :user AND usr.accessLevel = 'US'")
     Optional<Space> findByUserOwnerSpace(@Param("user") User user);
